Fix display of some parameters in a debug screen

+++ UI/DevScreens.cpp
@@ -1590,7 +1590,7 @@ void TouchTestScreen::DrawForeground(UIContext &dc) {
 		"g_dpi_scale_real: %0.3fx%0.3f\n"
 		"delta: %0.2f ms fps: %0.3f\n%s",
 #if PPSSPP_PLATFORM(ANDROID)
-		System_GetPropertyInt(SYSPROP_DISPLAY_XRES), System_GetPropertyInt(SYSPROP_DISPLAY_YRES),
+		(int)System_GetPropertyInt(SYSPROP_DISPLAY_XRES), (int)System_GetPropertyInt(SYSPROP_DISPLAY_YRES),
 #endif
 		g_display.dp_xres, g_display.dp_yres, g_display.pixel_xres, g_display.pixel_yres,
 		g_display.dpi, g_display.dpi_scale_x, g_display.dpi_scale_y,
